requirements.txt
# Production-grade face recognition
insightface==0.7.3
onnxruntime==1.18.1
onnx==1.16.1
numpy>=1.23
Pillow>=9.5
# Core Face Recognition Dependencies
deepface>=0.0.90
tensorflow==2.14.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# Image Processing
Pillow==10.0.1
numpy==1.24.3

# Additional Computer Vision Libraries
dlib==19.24.2
face-recognition==1.3.0

# System Dependencies
cmake==3.27.7

# Optional: GPU Support (uncomment if using CUDA)
# tensorflow-gpu==2.13.0

# Development Tools (optional)
# matplotlib==3.7.2
# scikit-learn==1.3.0